Cleanup page layouts
