+++ frontend/app-preview/src/PageLayout.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import classes from './PageLayout.module.css';
+import { Outlet } from 'react-router-dom';
+
+export const PageLayout = () => {
+  return (
+    <div className={classes.previewContainer}>
+      <Outlet />
+    </div>
+  );
+};
