+++ frontend/dashboard/routes/routes.tsx
@@ -5,7 +5,7 @@ import { Dashboard } from '../pages/Dashboard';
 import { CreateService } from '../pages/CreateService';
 
 export const routes = (
-  <Route path='/' element={<PageLayout />}>
+  <Route element={<PageLayout />}>
     <Route path='/:selectedContext?' element={<Dashboard />} />
     <Route path='/:selectedContext/new' element={<CreateService />} />
   </Route>
