--- frontend/packages/shared/src/PageLayout.module.css
@@ -1,10 +1,0 @@
-.container {
-  --toolbar-height: 80px;
-  --subtoolbar-height: 60px;
-  --header-border-bottom-width: 1px;
-  --header-height: calc(
-    var(--toolbar-height) + var(--subtoolbar-height) + var(--header-border-bottom-width)
-  );
-
-  height: calc(100vh - var(--header-height));
-}
