+++ frontend/packages/shared/src/PageLayout.tsx
@@ -1,11 +1,10 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import classes from './PageLayout.module.css';
 
 // TODO : https://github.com/Altinn/altinn-studio/issues/13271
 export const PageLayout = (): React.ReactNode => {
   return (
-    <div className={classes.container}>
+    <div>
       <Outlet />
     </div>
   );
